This should compile the stylesheets on deploy
